updated dataloader
