+++ keras_object_detection/utils/box_utils.py
@@ -66,6 +66,6 @@ def draw_boxes(image, boxes, categories):
             lineType=cv2.LINE_AA,
         )
         img = cv2.rectangle(
-            image, (_box[0], _box[1]), (_box[2], _box[3]), [30, 15, 30], 1
+            image, (_box[0], _box[1]), (_box[2], _box[3]), [0, 0, 0], 1
         )
     return img
